slog: Core functionality moved integration into module
The core functionality for secuure logging is now part of the corresponding module,
such that the code does not depend on any other part of syslog-ng. The commandline
utilities now have to link against the module library in addition to the core
syslog-ng library.
