+++ modules/secure-logging/Makefile.am
@@ -4,7 +4,8 @@ module_LTLIBRARIES					+=	\
 EXTRA_DIST += modules/secure-logging/CMakeLists.txt
 
 modules_secure_logging_libsecure_logging_la_SOURCES		=	\
-	modules/secure-logging/secure-logging.c
+	modules/secure-logging/secure-logging.c                         \
+	modules/secure-logging/slog.c
 modules_secure_logging_libsecure_logging_la_LIBADD		=	\
 	$(MODULE_DEPS_LIBS) $(OPENSSL_LIBS)
 modules_secure_logging_libsecure_logging_la_LDFLAGS		=	\
