+++ modules/secure-logging/slog.c
