+++ modules/secure-logging/slogimport/CMakeLists.txt
@@ -1,5 +1,5 @@
 
 add_executable(slogimport slogimport.c)
-target_link_libraries(slogimport eventlog)
+target_link_libraries(slogimport eventlog syslog-ng secure-logging)
 set_target_properties(slogimport PROPERTIES COMPILE_FLAGS "-Wno-pointer-sign")
 install(TARGETS slogimport RUNTIME DESTINATION bin)
