+++ modules/secure-logging/slogimport/slogimport.c
@@ -29,7 +29,7 @@
 #include "messages.h"
 
 
-#include "slog.h"
+#include "modules/secure-logging/slog.h"
 
 
 int main(int argc, char *argv[])
