+++ modules/secure-logging/slogverify/CMakeLists.txt
@@ -1,5 +1,5 @@
 
 add_executable(slogverify slogverify.c)
-target_link_libraries(slogverify eventlog)
+target_link_libraries(slogverify eventlog syslog-ng secure-logging)
 set_target_properties(slogverify PROPERTIES COMPILE_FLAGS "-Wno-pointer-sign")
 install(TARGETS slogverify RUNTIME DESTINATION bin)
